Refactor #8 - 스트림 메소드 이동
* 스트림 관련 기능 CustomStreamSupport 으로 이동
